+++ src/main/java/de/maxhenkel/voicechat/voice/client/Client.java
@@ -6,8 +6,12 @@ import de.maxhenkel.voicechat.config.ServerConfig;
 import de.maxhenkel.voicechat.events.ClientVoiceChatEvents;
 import de.maxhenkel.voicechat.voice.common.*;
 import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.network.chat.ComponentUtils;
+import net.minecraft.network.chat.HoverEvent;
+import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 
 import javax.annotation.Nullable;
@@ -135,12 +139,30 @@ public class Client extends Thread {
     }
 
     private void startMicThread() {
+        if (micThread != null) {
+            micThread.close();
+        }
         try {
             micThread = new MicThread(this);
             micThread.start();
         } catch (Exception e) {
-            Voicechat.LOGGER.error("Mic unavailable " + e);
+            Voicechat.LOGGER.error("Microphone unavailable: {}", e.getMessage());
+            sendPlayerError("messsage.voicechat.microphone_unavailable", e);
         }
+    }
+
+    public void sendPlayerError(String translationKey, Exception e) {
+        LocalPlayer player = Minecraft.getInstance().player;
+        if (player == null) {
+            return;
+        }
+        player.sendMessage(
+                ComponentUtils.wrapInSquareBrackets(new TextComponent(Voicechat.getModName()))
+                        .withStyle(ChatFormatting.GREEN)
+                        .append(" ")
+                        .append(new TranslatableComponent(translationKey).withStyle(ChatFormatting.RED))
+                        .withStyle(style -> style.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new TextComponent(e.getMessage()).withStyle(ChatFormatting.RED))))
+                , Util.NIL_UUID);
     }
 
     @Override