+++ changelog.md
@@ -1,1 +1,1 @@
-- Removed unnecessary accesswidener
+- Added microphone error message